requirements.txt
# local package
-e .

# internal requirements
tqdm
tensorflow==1.15.4
biopython==1.78
easy-parallel-py3==0.1.6.4
atom3-py3==0.1.4.4

# external requirements
click
Sphinx
coverage
awscli
flake8

# backwards compatibility
pathlib2